refactor(footer): modularize footer component with proper file structure and component separation
