+++ src/components/footer/Footer.scss
@@ -1,3 +1,4 @@
+// Footer.scss
 @use '../../styles/variables' as vars;
 @use '../../styles/mixins' as mixins;
 
@@ -23,219 +24,9 @@
     color: vars.$color-accent;
   }
 
-  &__form {
-    display: flex;
-    border: 1px solid vars.$color-border;
-    border-radius: 5px;
-    overflow: hidden;
-    margin-bottom: 1rem;
-
-    &:focus-within {
-      border-color: vars.$color-accent;
-    }
-  }
-
-  &__input {
-    border: none;
-    padding: 10px;
-    flex-grow: 1;
-    outline: none;
-    width: 100%;
-    background-color: vars.$color-background-light;
-    color: vars.$color-primary;
-
-    &:focus {
-      outline: 2px solid vars.$color-accent;
-    }
-  }
-
-  &__button {
-    border-left: 1px solid vars.$color-border;
-    border-radius: 0;
-    padding: 10px 20px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: vars.$gradient-button;
-    color: vars.$color-on-primary;
-    transition: vars.$transition-default;
-
-    &:hover {
-      color: vars.$color-background;
-      background-color: vars.$hover-color;
-    }
-
-    &:active {
-      background-color: vars.$color-primary;
-    }
-  }
-
-  &__message {
-    text-align: center;
-    color: vars.$color-accent;
-    margin-bottom: 1rem;
-  }
-
-  &__info {
-    margin-bottom: 1rem;
-    text-align: center;
-
-    a {
-      color: vars.$color-secondary;
-      text-decoration: none;
-
-      &:hover {
-        color: vars.$hover-color;
-        text-decoration: underline;
-      }
-    }
-  }
-
-  &__whatsapp-link {
-    margin-left: 10px;
-    color: vars.$color-primary;
-
-    &:hover {
-      color: vars.$color-accent;
-    }
-  }
-
-  &__whatsapp-icon {
-    font-size: 1.25rem;
-  }
-
-  &__contact {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0.625rem;
-    margin-bottom: 1rem;
-
-    a {
-      color: vars.$color-secondary;
-      text-decoration: none;
-
-      &:hover {
-        color: vars.$hover-color;
-        text-decoration: underline;
-      }
-    }
-  }
-
-  &__social {
-    display: flex;
-    gap: 0.75rem;
-
-    a {
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      min-width: 48px;
-      min-height: 48px;
-      text-decoration: none;
-      color: vars.$color-secondary;
-
-      &:hover {
-        color: vars.$hover-color;
-        text-decoration: none;
-      }
-    }
-
-    i {
-      font-size: 1.25rem;
-      line-height: 1;
-    }
-  }
-
-  &__legal {
-    display: flex;
-    justify-content: center;
-    gap: 0.5rem;
-    margin-bottom: 1rem;
-
-    a {
-      color: vars.$color-secondary;
-      text-decoration: none;
-
-      &:hover {
-        color: vars.$hover-color;
-        text-decoration: underline;
-      }
-    }
-  }
-
   &__copyright {
     text-align: center;
     margin-bottom: 1rem;
     color: vars.$color-secondary;
-  }
-}
-
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
-
-@media (min-width: 768px) {
-  .footer {
-    padding: 1.875rem;
-    font-size: 1rem;
-
-    &__form {
-      max-width: 500px;
-      margin: 0 auto 1rem;
-    }
-
-    &__contact {
-      flex-direction: row;
-      justify-content: center;
-    }
-  }
-}
-
-@media (min-width: 1024px) {
-  .footer {
-    padding: 1.875rem 0;
-    font-size: 1.125rem;
-
-    &__container {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      text-align: center;
-    }
-
-    &__title {
-      margin-bottom: 1rem;
-    }
-
-    &__form {
-      margin-bottom: 1.5rem;
-    }
-
-    &__info {
-      margin-bottom: 1.5rem;
-    }
-
-    &__contact {
-      align-items: center;
-      margin-bottom: 1.5rem;
-    }
-
-    &__legal {
-      margin-bottom: 1rem;
-    }
-
-    &__copyright {
-      margin-bottom: 1.5rem;
-    }
   }
 }
