+++ src/components/footer/components/ContactInfo.scss
@@ -1,0 +1,31 @@
+// components/ContactInfo.scss
+@use '../../../styles/variables' as vars;
+@use '../../../styles/mixins' as mixins;
+
+.contact-info {
+  margin-bottom: 1rem;
+  text-align: center;
+
+  a {
+    color: vars.$color-secondary;
+    text-decoration: none;
+
+    &:hover {
+      color: vars.$hover-color;
+      text-decoration: underline;
+    }
+  }
+
+  &__whatsapp-link {
+    margin-left: 10px;
+    color: vars.$color-primary;
+
+    &:hover {
+      color: vars.$color-accent;
+    }
+  }
+
+  &__whatsapp-icon {
+    font-size: 1.25rem;
+  }
+}
