+++ src/components/footer/components/LegalSection.scss
@@ -1,0 +1,20 @@
+// components/LegalSection.scss
+@use '../../../styles/variables' as vars;
+@use '../../../styles/mixins' as mixins;
+
+.legal-section {
+  display: flex;
+  justify-content: center;
+  gap: 0.5rem;
+  margin-bottom: 1rem;
+
+  &__link {
+    color: vars.$color-secondary;
+    text-decoration: none;
+
+    &:hover {
+      color: vars.$hover-color;
+      text-decoration: underline;
+    }
+  }
+}
