+++ src/components/footer/components/SocialLinks.scss
@@ -1,0 +1,30 @@
+// components/SocialLinks.scss
+@use '../../../styles/variables' as vars;
+@use '../../../styles/mixins' as mixins;
+
+.social-links {
+  display: flex;
+  gap: 0.75rem;
+  justify-content: center;
+  margin-bottom: 1rem;
+
+  &__link {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 48px;
+    min-height: 48px;
+    text-decoration: none;
+    color: vars.$color-secondary;
+
+    &:hover {
+      color: vars.$hover-color;
+      text-decoration: none;
+    }
+
+    i {
+      font-size: 1.25rem;
+      line-height: 1;
+    }
+  }
+}
